beta 1.1
small change

+++ src/main/java/com/intern/project/POJO/Score.java
@@ -25,11 +25,11 @@ public class Score {
 
 	
 	
-	public int getID() {
+	public long getID() {
 		return ID;
 	}
 
-	public void setID(int iD) {
+	public void setID(long iD) {
 		ID = iD;
 	}
